Finished logic on si page
